requirements.txt
Flask
python-dotenv
Werkzeug
requests
langchain
langchain-community
langchain-pinecone
langchain-huggingface
openai
pinecone-client
googlemaps
huggingface-hub
pypdf
gunicorn
flask-sqlalchemy
transformers
torch
flask-cors
speechrecognition
together
apscheduler